\name{mm_metabolic}
\alias{mm_metabolic}
\docType{data}
\title{Metabolic Pathways Data for Mouse}
\description{
Gene Set Knowledgebase (GSKB) is a comprehensive knowledgebase for pathway 
analysis in mouse. This object contains metabolic pathways based on KEGG, EHMN,
MouseCyc and REACTOME.  This data can be used by programs such as GSEA and 
pGSEA. 
}
\usage{data(mm_metabolic)}
\references{Liming Lai, Jason Hennessey, Yuguang Ban, Wei Wang, Jianli Qi, 
Gaixin Jiang, Valerie Bares, and Steven Xijin Ge, GSKB: A comprehensive 
knowledgebase for pathway analysis in mouse. Database: The journal of 
biological databases and curation (in revision).}
\keyword{ datasets }
\value{A list is returned, where each list element is a gene set.}