\name{mm_TF}
\alias{mm_TF}
\docType{data}
\title{Transcription Factor Target Genes Data for Mouse}
\description{
Gene Set Knowledgebase (GSKB) is a comprehensive knowledgebase for pathway 
analysis in mouse. This object contains gene sets corresponding to target genes
of transcription factors. Some gene sets are based on experimentally confirmed 
data, while others are based on predictions. These are based on TFacts, TRED, 
CircuitsDB, and TRANSFAC.
}
\usage{data(mm_TF)}
\references{Liming Lai, Jason Hennessey, Yuguang Ban, Wei Wang, Jianli Qi, 
Gaixin Jiang, Valerie Bares, and Steven Xijin Ge, GSKB: A comprehensive 
knowledgebase for pathway analysis in mouse. Database: The journal of 
biological databases and curation (in revision).}
\keyword{ datasets }
\value{A list is returned, where each list element is a gene set.}